+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Make filepath argument optional: Issue [#170](https://github.com/nowarp/misti/issues/170)
+
 ## [0.4.0] - 2024-10-08
 
 ### Added