+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.1] - 2024-10-12
+
 ### Changed
 - Make the filepath argument optional: Issue [#170](https://github.com/nowarp/misti/issues/170)